requirements.txt
line-bot-sdk==3.11.0
python-dotenv==0.21.1
Flask==2.2.2
opencc-python-reimplemented==0.1.4
beautifulsoup4==4.11.2
youtube-transcript-api==0.6.2
pymongo==4.3.3
aiohttp ==3.9.5
Werkzeug==2.2.2
flask-ngrok
openai
pandas
pandas_ta
yfinance
numpy
mkdocs-techdocs-core